access-esdb/requirements_backup.txt
aniso8601==9.0.1
attrs==23.1.0
certifi==2022.12.7
charset-normalizer 3.1.0
click==8.1.3
elastic-transport==8.4.0
elasticsearch==8.7.0
Flask==2.0.3
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.9
Flask-Testing==0.8.1
idna==3.4
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
markup==0.2
MarkupSafe==2.1.2
numpy==1.24.3
opa-python==0.0.8
opa-python-client==1.3.4
packaging==23.1
pandas==2.0.1
pip==23.0.1
pluggy==1.0.0
pyopa==0.8.2
pyrsistent==0.19.3
pytest==7.3.1
pytest-flask==1.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
setuptools==66.0.0
six==1.16.0
types-requests==2.29.0.0
types-urllib3==1.26.25.12
tzdata==2023.3
urllib3==1.26.15
user-agent==0.1.10
Werkzeug==2.3.3
wheel==0.38.4
